app/requirements.txt
# Web framework
flask==2.0.1
gunicorn==20.1.0

# Database
mysql-connector-python==8.0.26
SQLAlchemy==1.4.23

# Security
cryptography==3.4.8
passlib==1.7.4
flask-login==0.5.0

# Utilities
python-dotenv==0.19.0
requests==2.26.0

# Testing
pytest==6.2.5
pytest-cov==2.12.1
